Make scala.deriving a package instead of an object.
The fundamental `Mirror` trait receives its own file. The helpers
`ArrayProduct`, `EmptyProduct` and `productElement` are put
together in a `Helpers.scala` file.
